Реализует слайдер
